[#3679] Move IndicatorPeriodModelTestCase to separate test module
