+++ akvo/rsr/tests/models/test_indicator.py
@@ -6,13 +6,10 @@
 
 import datetime
 
-from unittest import TestCase
-
 from django.contrib.auth import get_user_model
 from django.core.exceptions import ValidationError
 
-from akvo.rsr.models import (Indicator, IndicatorPeriod, IndicatorPeriodData, Project, Result,
-                             RelatedProject)
+from akvo.rsr.models import Indicator, IndicatorPeriod, Project, Result, RelatedProject
 from akvo.rsr.tests.base import BaseTestCase
 from akvo.rsr.models.result.utils import QUALITATIVE
 
@@ -127,114 +124,3 @@ class IndicatorModelTestCase(BaseTestCase):
         child_indicator = parent_indicator.child_indicators.first()
         self.assertEqual(QUALITATIVE, parent_indicator.type)
         self.assertEqual(parent_indicator.type, child_indicator.type)
-
-
-class IndicatorPeriodModelTestCase(TestCase):
-    """Tests for the indicator period model"""
-
-    def setUp(self):
-        # Clear all projects, users since some tests may not tear down!
-        self.tearDown()
-
-        # Setup a project with results framework and a user
-        self.project = Project.objects.create(title="Test project 1")
-        self.result = Result.objects.create(project=self.project, title='Test Result')
-        self.indicator = Indicator.objects.create(result=self.result, title='Test Indicator')
-        self.period = IndicatorPeriod.objects.create(indicator=self.indicator,
-                                                     actual_comment='initial actual comment')
-        self.user = User.objects.create(username='user1@com.com', email='user1@com.com')
-
-    def tearDown(self):
-        Project.objects.all().delete()
-        User.objects.all().delete()
-
-    def test_unapproved_period_data_does_not_update_actual_comment(self):
-        """Unapproved IndicatorPeriodData doesn't update IndicatorPeriod.actual_comment."""
-
-        # Given
-        period = self.period
-        user = self.user
-
-        # When
-        data = IndicatorPeriodData.objects.create(text='period data comment',
-                                                  period=period,
-                                                  user=user)
-
-        # Then
-        period = IndicatorPeriod.objects.get(id=period.id)
-        self.assertNotIn(data.text, period.actual_comment)
-
-    def test_period_data_updates_actual_comment(self):
-        """Adding IndicatorPeriodData updates IndicatorPeriod.actual_comment."""
-
-        # Given
-        period = self.period
-        user = self.user
-
-        # When
-        data = IndicatorPeriodData.objects.create(text='period data comment',
-                                                  period=period,
-                                                  user=user,
-                                                  status=IndicatorPeriodData.STATUS_APPROVED_CODE)
-
-        # Then
-        period = IndicatorPeriod.objects.get(id=period.id)
-        self.assertIn(data.text, period.actual_comment)
-
-    def test_period_data_non_ascii_updates_actual_comment(self):
-        """Test that adding IndicatorPeriodData with unicode text works."""
-
-        # Given
-        period = self.period
-        user = self.user
-
-        # When
-        data = IndicatorPeriodData.objects.create(text=u'perïöd data comment',
-                                                  period=period,
-                                                  user=user,
-                                                  status=IndicatorPeriodData.STATUS_APPROVED_CODE)
-
-        # Then
-        period = IndicatorPeriod.objects.get(id=period.id)
-        self.assertIn(data.text, period.actual_comment)
-
-    def test_multiple_period_data_updates_actual_comment(self):
-        """Test that adding multiple IndicatorPeriodData updates actual_comment."""
-
-        # Given
-        period = self.period
-        user = self.user
-
-        # When
-        data_1 = IndicatorPeriodData.objects.create(text=u'perïöd data_1 comment',
-                                                    period=period,
-                                                    user=user,
-                                                    status=IndicatorPeriodData.STATUS_APPROVED_CODE)
-        data_2 = IndicatorPeriodData.objects.create(text=u'perïöd data_2 comment',
-                                                    period=period,
-                                                    user=user,
-                                                    status=IndicatorPeriodData.STATUS_APPROVED_CODE)
-
-        # Then
-        period = IndicatorPeriod.objects.get(id=period.id)
-        self.assertIn(data_2.text, period.actual_comment)
-        self.assertIn(data_1.text, period.actual_comment)
-        # newer update's text appears before older one's
-        self.assertLess(period.actual_comment.index(data_2.text), period.actual_comment.index(data_1.text))
-
-    def test_period_data_deletion_updates_actual_comment(self):
-
-        # Given
-        period = self.period
-        user = self.user
-        data = IndicatorPeriodData.objects.create(text='period data comment',
-                                                  period=period,
-                                                  user=user,
-                                                  status=IndicatorPeriodData.STATUS_APPROVED_CODE)
-
-        # When
-        data.delete()
-
-        # Then
-        period = IndicatorPeriod.objects.get(id=period.id)
-        self.assertNotIn(data.text, period.actual_comment)
